refactor!(NODE-3427): remove md5 hashing from GridFS API
